Carousel working for now. Will need to continue to adjust sizing
