+++ src/components/Nav/Nav.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Nav = () => {
+  return (
+    <nav>
+      <ul>
+        <li>Locations</li>
+        <li>Podcast Archive</li>
+        <li>Coffee Menu</li>
+        <li>Contact</li>
+      </ul>
+    </nav>
+  )
+}
+
+export default Nav;
