+++ src/styles/_App.scss
@@ -1,0 +1,30 @@
+* {
+  box-sizing: border-box;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 16px;
+}
+
+.App {
+  display: grid;
+  grid-template-areas: 
+    'nav carousel'
+    'nav locations'
+    'nav locations'
+    'nav locations';
+  grid-template-columns: 15% 1fr;
+  height: 100vh;
+}
+
+.slider {
+  grid-area: carousel
+}
+
+.slider-slide > img {
+  height: 500px;
+}
+
+.locations {
+  grid-area: locations;
+  height: 50vh;
+  background: blue;
+}
