+++ src/styles/_Nav.scss
@@ -1,0 +1,11 @@
+nav {
+  grid-area: nav;
+  background: black;
+  color: white;
+  width: 100%;
+
+  ul {
+    list-style: none;
+    padding: 0;
+  }
+}
