+++ src/styles/main.scss
@@ -1,0 +1,2 @@
+@import "App";
+@import "Nav";
